Extract ModelStub into new file to reduce code duplicate.
